requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
appdirs==1.4.4
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
behave==1.2.6
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==40.0.2
cssselect==1.2.0
exceptiongroup==1.1.1
execnet==1.9.0
fasteners==0.18
filelock==3.12.0
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
jaraco.classes==3.2.3
keyring==23.13.1
markdown-it-py==2.2.0
mdurl==0.1.2
more-itertools==9.1.0
outcome==1.2.0
packaging==23.1
parameterized==0.9.0
parse==1.19.0
parse-type==0.6.0
pdbp==1.4.0
platformdirs==3.5.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyee==7.0.4
Pygments==2.15.1
pynose==1.4.4
pyOpenSSL==23.1.1
pyotp==2.8.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyppeteer2==0.2.2
PySocks==1.7.1
pytest==7.3.1
pytest-forked==1.6.0
pytest-html==2.0.1
pytest-metadata==2.0.4
pytest-ordering==0.6
pytest-rerunfailures==11.1.2
pytest-xdist==3.2.1
python-dotenv==1.0.0
PyYAML==6.0
requests==2.30.0
requests-file==1.5.1
requests-toolbelt==1.0.0
rich==13.3.5
sbvirtualdisplay==1.2.0
selenium==4.9.1
selenium-requests==2.0.3
seleniumbase==4.14.8
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tabcompleter==1.2.0
tldextract==3.4.1
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.15
webdriver-manager==3.8.6
websockets==8.1
wsproto==1.2.0
zipp==3.15.0
